+++ haoskiosk/CHANGELOG.md
@@ -1,8 +1,10 @@
 # Changelog
 
-## v0.9.8 – May 2025
+## v0.9.8 – June 2025
 
 - Added ability to set browser theme and sidebar behavior
+- Added <Control-r> binding to reload browser screen
+- Refresh screen automatically when returning from screen blank
 - Improved input validation and error handling
 - Tightened up code
 - Updated documentation